HTML Part 2 Classwork
I tested the file in  https://validator.w3.org/ and fix a couple of errors. I couldn't figure what was wrong with the <spam> cannot be in <label> tag however. I did some research and it said that sometimes the validator gets confused because a tag above was enclosed.  I will upload an update once I figure out if this is what is happening or if I made a mistake in the code itself.